pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pybatgym"
version = "0.1.0"
description = "A Gymnasium-compatible RL environment for HPC scheduling on BatSim"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "PyBatGym Team" },
]
dependencies = [
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
rl = [
    "stable-baselines3>=2.0.0",
    "sb3-contrib>=2.0.0",
]
all = ["pybatgym[dev,rl]"]

[tool.setuptools.packages.find]
include = ["pybatgym*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
